Move all localized classes to Localized module.
