+++ lib/twitter_cldr/core_ext.rb
@@ -3,15 +3,10 @@
 # Copyright 2012 Twitter, Inc
 # http://www.apache.org/licenses/LICENSE-2.0
 
-require 'twitter_cldr/core_ext/localized_object'
-require 'twitter_cldr/core_ext/calendars/datetime'
-require 'twitter_cldr/core_ext/calendars/date'
-require 'twitter_cldr/core_ext/calendars/time'
-require 'twitter_cldr/core_ext/calendars/timespan'
-require 'twitter_cldr/core_ext/numbers/localized_number'
-require 'twitter_cldr/core_ext/numbers/bignum'
-require 'twitter_cldr/core_ext/numbers/fixnum'
-require 'twitter_cldr/core_ext/numbers/float'
-require 'twitter_cldr/core_ext/symbol'
-require 'twitter_cldr/core_ext/string'
-require 'twitter_cldr/core_ext/array'
+[Bignum, Fixnum, Float].each do |klass|
+  TwitterCldr::Localized::LocalizedNumber.localize(klass)
+end
+
+[Array, Date, DateTime, String, Symbol, Time].each do |klass|
+  TwitterCldr::Localized::const_get("Localized#{klass}").localize(klass)
+end
