--- lib/twitter_cldr/core_ext/numbers/float.rb
@@ -1,6 +1,0 @@
-# encoding: UTF-8
-
-# Copyright 2012 Twitter, Inc
-# http://www.apache.org/licenses/LICENSE-2.0
-
-TwitterCldr::LocalizedNumber.localize(Float)
