--- lib/twitter_cldr/core_ext/string.rb
@@ -1,41 +1,0 @@
-# encoding: UTF-8
-
-# Copyright 2012 Twitter, Inc
-# http://www.apache.org/licenses/LICENSE-2.0
-
-module TwitterCldr
-
-  class LocalizedString < LocalizedObject
-
-    # Uses wrapped string object as a format specification and returns the result of applying it to +args+ (see
-    # +TwitterCldr::Utils.interpolate+ method for interpolation syntax).
-    #
-    # If +args+ is a Hash than pluralization is performed before interpolation (see +PluralFormatter+ class for
-    # pluralization specification).
-    #
-    def %(args)
-      pluralized = args.is_a?(Hash) ? @formatter.format(@base_obj, args) : @base_obj
-      TwitterCldr::Utils.interpolate(pluralized, args)
-    end
-
-    def formatter_const
-      TwitterCldr::Formatters::PluralFormatter
-    end
-
-    def normalize(options = {})
-      TwitterCldr::Normalization.normalize(@base_obj, options).localize(@locale)
-    end
-
-    def code_points
-      TwitterCldr::Utils::CodePoints.from_string(@base_obj)
-    end
-
-    def to_s
-      @base_obj.dup
-    end
-
-  end
-
-  LocalizedString.localize(String)
-
-end
