+++ lib/twitter_cldr/localized/localized_time.rb
@@ -1,0 +1,23 @@
+# encoding: UTF-8
+
+# Copyright 2012 Twitter, Inc
+# http://www.apache.org/licenses/LICENSE-2.0
+
+module TwitterCldr
+  module Localized
+
+    class LocalizedTime < LocalizedDateTime
+      def to_datetime(date)
+        date_obj = date.is_a?(LocalizedDate) ? date.base_obj : date
+        LocalizedDateTime.new(DateTime.parse("#{date_obj.strftime("%Y-%m-%d")}T#{@base_obj.strftime("%H:%M:%S%z")}"), @locale, :calendar_type => @calendar_type)
+      end
+
+      protected
+
+      def formatter_const
+        TwitterCldr::Formatters::TimeFormatter
+      end
+    end
+
+  end
+end
