+++ lib/twitter_cldr/localized/localized_timespan.rb
@@ -1,0 +1,26 @@
+# encoding: UTF-8
+
+# Copyright 2012 Twitter, Inc
+# http://www.apache.org/licenses/LICENSE-2.0
+
+module TwitterCldr
+  module Localized
+
+    class LocalizedTimespan < LocalizedObject
+      def initialize(seconds, options = {})
+        super(seconds, options[:locale], options)
+      end
+
+      def to_s(options = {})
+        @formatter.format(@base_obj, options)
+      end
+
+      protected
+
+      def formatter_const
+        TwitterCldr::Formatters::TimespanFormatter
+      end
+    end
+
+  end
+end
